refactor: consolidate useLocalStorage and useSessionStorage into useStorage for improved code organization
